+++ linux/cc/WindowX11.cc
@@ -42,7 +42,7 @@ void WindowX11::setClass(const std::string& name, const std::string& appClass) {
     }
 }
 
-void WindowX11::setIconData(int width, int height, const unsigned char* argb) {
+void WindowX11::setIconPixels(int width, int height, const unsigned char* argb) {
     size_t size = width * height;
     size_t count = size + 2;
     std::unique_ptr<long[]> buffer{new long[count]};
@@ -609,13 +609,12 @@ extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_WindowX11__1nSetCl
     instance->setClass(bytesToString(env, name), bytesToString(env, appClass));
 }
 
-extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_WindowX11__1nSetIconData
-        (JNIEnv* env, jobject obj, jint width, jint height, jbyteArray data) {
+extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_WindowX11__1nSetIconPixels
+        (JNIEnv* env, jobject obj, jint width, jint height, jbyteArray pixelsArr) {
     jwm::WindowX11* instance = reinterpret_cast<jwm::WindowX11*>(jwm::classes::Native::fromJava(env, obj));
-
-    jbyte* bytes = env->GetByteArrayElements(data, nullptr);
-    instance->setIconData(width, height, reinterpret_cast<const unsigned char*>(bytes));
-    env->ReleaseByteArrayElements(data, bytes, 0);
+    jbyte* pixels = env->GetByteArrayElements(pixelsArr, nullptr);
+    instance->setIconPixels(width, height, reinterpret_cast<const unsigned char*>(pixels));
+    env->ReleaseByteArrayElements(pixelsArr, pixels, 0);
 }
 
 extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_WindowX11__1nSetTitlebarVisible