+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 
 - Bumped types in pom.xml to 0.2.0
+- Windows, Linux: Window::setIconPixels (previously setIconData on Linux) #310 #135 via @chirontt 
 
 # 0.4.25 - Jun 16, 2026